Minor: restructuring of code and cleanup
